+++ src/main/java/fr/maxlego08/menu/command/commands/CommandMenu.java
@@ -9,6 +9,7 @@ import fr.maxlego08.menu.command.commands.players.CommandMenuPlayers;
 import fr.maxlego08.menu.command.commands.reload.CommandMenuReload;
 import fr.maxlego08.menu.command.commands.website.CommandMenuConnect;
 import fr.maxlego08.menu.command.commands.website.CommandMenuDisconnect;
+import fr.maxlego08.menu.command.commands.website.CommandMenuUnlink;
 import fr.maxlego08.menu.command.commands.website.CommandMenuDownload;
 import fr.maxlego08.menu.command.commands.website.CommandMenuInventories;
 import fr.maxlego08.menu.command.commands.website.CommandMenuLogin;
@@ -43,6 +44,7 @@ public class CommandMenu extends VCommand {
         this.addSubCommand(new CommandMenuLogin(plugin));
         this.addSubCommand(new CommandMenuConnect(plugin));
         this.addSubCommand(new CommandMenuDisconnect(plugin));
+        this.addSubCommand(new CommandMenuUnlink(plugin));
         this.addSubCommand(new CommandMenuInventories(plugin));
         // this.addSubCommand(new CommandMenuMarketplace(plugin));